SwitchCaseExamples
This project is prepared to learn the use of switch-case structure in C# language. The switch-case structure allows different code blocks to be executed according to the value of a given variable and provides a more organised structure instead of a large number of if-else statements. In this example, different operations are performed depending on a number or character.